linux: Move `types.h` API to `src/new`
Link: https://github.com/torvalds/linux/blob/dc59e4fea9d83f03bad6bddf3fa2e52491777482/include/uapi/linux/types.h
Link: https://github.com/torvalds/linux/blob/dc59e4fea9d83f03bad6bddf3fa2e52491777482/include/uapi/asm-generic/types.h
Link: https://github.com/torvalds/linux/blob/dc59e4fea9d83f03bad6bddf3fa2e52491777482/include/uapi/asm-generic/int-ll64.h
Link: https://cs.android.com/android/platform/superproject/main/+/main:bionic/libc/kernel/uapi/linux/types.h;drc=180edefbd287c39caeb9d48784a9a10ac35f3636
Link: https://cs.android.com/android/platform/superproject/main/+/main:bionic/libc/kernel/uapi/asm-generic/types.h;drc=180edefbd287c39caeb9d48784a9a10ac35f3636
Link: https://cs.android.com/android/platform/superproject/main/+/main:bionic/libc/kernel/uapi/asm-generic/int-ll64.h;drc=180edefbd287c39caeb9d48784a9a10ac35f3636

(backport <https://github.com/rust-lang/libc/pull/5233>)
(cherry picked from commit 483795464d6fbbd5fe07d59013f7c9e2bb46ec82)
