+++ src/new/bionic_libc/kernel_uapi/linux/mod.rs
@@ -1,0 +1,5 @@
+//! Directory: `bionic/libc/kernel/uapi/linux/`
+//!
+//! <https://cs.android.com/android/platform/superproject/main/+/main:bionic/libc/kernel/uapi/linux/>
+
+pub(crate) mod types;
