+++ src/new/bionic_libc/kernel_uapi/mod.rs
@@ -1,0 +1,3 @@
+//! Directory: `bionic/libc/kernel/uapi`
+
+pub(crate) mod linux;
