+++ src/new/linux_uapi/linux/mod.rs
@@ -13,3 +13,4 @@ pub(crate) mod netlink;
 pub(crate) mod pidfd;
 pub(crate) mod sctp;
 pub(crate) mod tls;
+pub(crate) mod types;
