+++ src/unix/linux_like/android/b32/mod.rs
@@ -8,8 +8,6 @@ pub type off64_t = c_longlong;
 pub type sigset_t = c_ulong;
 pub type socklen_t = i32;
 pub type time64_t = i64;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 s! {
     // FIXME(1.0): This should not implement `PartialEq`
