+++ src/unix/linux_like/android/b64/aarch64/mod.rs
@@ -2,8 +2,6 @@ use crate::off64_t;
 use crate::prelude::*;
 
 pub type wchar_t = u32;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 s! {
     pub struct stat {
