+++ src/unix/linux_like/android/b64/riscv64/mod.rs
@@ -3,8 +3,6 @@ use crate::prelude::*;
 
 pub type wchar_t = u32;
 pub type greg_t = i64;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 s! {
     pub struct stat {
