+++ src/unix/linux_like/linux/gnu/b64/aarch64/mod.rs
@@ -10,8 +10,6 @@ pub type wchar_t = u32;
 pub type nlink_t = u32;
 pub type blksize_t = i32;
 pub type suseconds_t = i64;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 s! {
     // FIXME(1.0): This should not implement `PartialEq`
