+++ src/unix/linux_like/linux/gnu/b64/mips64/mod.rs
@@ -9,8 +9,6 @@ pub type blksize_t = i64;
 pub type nlink_t = u64;
 pub type suseconds_t = i64;
 pub type wchar_t = i32;
-pub type __u64 = c_ulong;
-pub type __s64 = c_long;
 
 s! {
     pub struct stat {
