+++ src/unix/linux_like/linux/gnu/b64/powerpc64/mod.rs
@@ -11,8 +11,6 @@ pub type wchar_t = i32;
 pub type nlink_t = u64;
 pub type blksize_t = i64;
 pub type suseconds_t = i64;
-pub type __u64 = c_ulong;
-pub type __s64 = c_long;
 pub type gregset_t = [c_ulong; __NGREG];
 pub type fpregset_t = [c_ulong; __NFPREG];
