+++ src/unix/linux_like/linux/gnu/b64/riscv64/mod.rs
@@ -14,8 +14,6 @@ pub type blksize_t = c_int;
 pub type fsblkcnt64_t = c_ulong;
 pub type fsfilcnt64_t = c_ulong;
 pub type suseconds_t = i64;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 s! {
     pub struct pthread_attr_t {
