+++ src/unix/linux_like/linux/gnu/b64/s390x.rs
@@ -12,8 +12,6 @@ pub type nlink_t = u64;
 pub type suseconds_t = i64;
 pub type wchar_t = i32;
 pub type greg_t = u64;
-pub type __u64 = u64;
-pub type __s64 = i64;
 
 s! {
     // FIXME(1.0): This should not implement `PartialEq`
