+++ src/unix/linux_like/linux/mod.rs
@@ -1,6 +1,12 @@
 //! Linux-specific definitions for linux-like values
 use crate::prelude::*;
 use crate::{
+    __s16,
+    __s32,
+    __u16,
+    __u32,
+    __u64,
+    __u8,
     sock_filter,
     _IO,
     _IOR,
@@ -25,12 +31,6 @@ pub type pthread_once_t = c_int;
 pub type pthread_spinlock_t = c_int;
 pub type __kernel_fsid_t = __c_anonymous__kernel_fsid_t;
 pub type __kernel_clockid_t = c_int;
-
-pub type __u8 = c_uchar;
-pub type __u16 = c_ushort;
-pub type __s16 = c_short;
-pub type __u32 = c_uint;
-pub type __s32 = c_int;
 
 pub type eventfd_t = u64;
