+++ src/unix/linux_like/linux/musl/b64/aarch64/mod.rs
@@ -1,8 +1,6 @@
 use crate::off_t;
 use crate::prelude::*;
 
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 pub type wchar_t = u32;
 pub type nlink_t = u32;
 pub type blksize_t = c_int;
