+++ src/unix/linux_like/linux/musl/b64/mips64.rs
@@ -2,8 +2,6 @@ use crate::off_t;
 use crate::prelude::*;
 
 pub type wchar_t = i32;
-pub type __u64 = c_ulong;
-pub type __s64 = c_long;
 pub type nlink_t = c_uint;
 pub type blksize_t = i64;
