+++ src/unix/linux_like/linux/musl/b64/powerpc64.rs
@@ -2,8 +2,6 @@ use crate::off_t;
 use crate::prelude::*;
 
 pub type wchar_t = i32;
-pub type __u64 = c_ulong;
-pub type __s64 = c_long;
 pub type nlink_t = u64;
 pub type blksize_t = c_long;
