+++ src/unix/linux_like/linux/musl/b64/riscv64/mod.rs
@@ -7,8 +7,6 @@ pub type wchar_t = c_int;
 
 pub type nlink_t = c_uint;
 pub type blksize_t = c_int;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 pub type stat64 = stat;
