+++ src/unix/linux_like/linux/musl/b64/s390x.rs
@@ -5,8 +5,6 @@ pub type blksize_t = i64;
 pub type nlink_t = u64;
 pub type wchar_t = i32;
 pub type greg_t = u64;
-pub type __u64 = u64;
-pub type __s64 = i64;
 pub type statfs64 = statfs;
 pub type stat64 = stat;
