+++ src/unix/linux_like/linux/musl/b64/wasm32/mod.rs
@@ -6,8 +6,6 @@ use crate::prelude::*;
 pub type wchar_t = i32;
 pub type nlink_t = u64;
 pub type blksize_t = c_long;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 pub type stat64 = stat;
