+++ src/unix/linux_like/linux/uclibc/arm/mod.rs
@@ -16,8 +16,6 @@ pub type blkcnt_t = c_long;
 
 pub type fsblkcnt64_t = u64;
 pub type fsfilcnt64_t = u64;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 
 s! {
     pub struct cmsghdr {
