+++ src/unix/linux_like/linux/uclibc/mips/mips32/mod.rs
@@ -10,8 +10,6 @@ pub type blksize_t = i32;
 pub type nlink_t = u32;
 pub type fsblkcnt_t = c_ulong;
 pub type fsfilcnt_t = c_ulong;
-pub type __u64 = c_ulonglong;
-pub type __s64 = c_longlong;
 pub type fsblkcnt64_t = u64;
 pub type fsfilcnt64_t = u64;
